Added Reports

+++ Models/preprocess.py
@@ -4,7 +4,7 @@ def separate(X):
     X.set_index('Tour_ID', inplace=True)
     
     y=[]
-    num=[col for col in X.select_dtypes('int','float').columns]
+    num=[col for col in X.select_dtypes('int64','float64').columns]
     
     cat=[]
     text=[]
